Added Support for Views and Models
The constructor of the Controller now takes an array as its first argument. In that array, you can provide a viewsDir and modelsDir with paths to their relevant folders/directories. You do not need to create a constructor for your Controller, as calling the constructor during instantiation will call the parent Controller constructor and pass the args for you.

Render a view by calling the view() method with a string as the view name. Same goes for models.